hal: rename aux to util
