+++ wgpu-core/src/device/mod.rs
@@ -2144,7 +2144,7 @@ impl<A: HalApi> Device<A> {
                     }
                     hal::PipelineError::EntryPoint(stage) => {
                         pipeline::CreateRenderPipelineError::Internal {
-                            stage: hal::aux::map_naga_stage(stage),
+                            stage: hal::util::map_naga_stage(stage),
                             error: EP_FAILURE.to_string(),
                         }
                     }
