+++ wgpu-hal/src/metal/device.rs
@@ -5,7 +5,7 @@ use std::{
 };
 
 use super::conv;
-use crate::aux::map_naga_stage;
+use crate::util::map_naga_stage;
 
 type DeviceResult<T> = Result<T, crate::DeviceError>;
